refactor(ui): split SpeechCommandsMixin into three cohesive mixins (CQ-1)
main_frame_speech.py drops 2608 -> 462 lines (budget was a stale 2931) by
moving two cohesive surfaces verbatim into new mixin modules, both composed
onto MainFrame alongside the core:

- SpeechDownloadsMixin (main_frame_speech_downloads.py, 1784): the model /
  optional-component acquisition surface -- Speech Models manager,
  HuggingFace token prompt, Optional Components hub (install/remove/test/
  manage), guided offline-speech setup, and every download_* acquisition
  (ffmpeg, faster-whisper, vosk, DECtalk, Piper, Node, eSpeak, Kokoro,
  braille pack, MathCAT, PDF-OCR, Pandoc, audio extras) plus per-model
  download/remove/test with progress.
- VoiceInteractionMixin (main_frame_speech_voice.py, 724): the offline
  voice surface -- dictation, voice commands (#663, allowlist-dispatched),
  conversation mode (Hey QUILL Phase 2: VAD + spoken cues + Ask Quill
  routing), and the wake-word listener (Phase 3).

The core file keeps the provider/registry accessors, model prewarm,
transcription, caption generation, microphone selection, and command
registration. Voice safety law is unchanged (off by default, Safe-Mode
gated, SAFE_TOOL_IDS-only dispatch).

Budgets ratcheted (three new entries; +4 on main_frame.py for the mixin
wiring). Dialog inventory regenerated (keys moved 1:1). Source-contract
tests that grep main_frame_speech.py now scan all three main_frame_speech*.py
files. No behavior change; full UI unit suite green (the only failures were
the known media/timer flakes that pass in isolation).
