+++ quill/tools/module_size_budgets.json
@@ -449,7 +449,8 @@
     "_rebaseline_2026_07_08_speech_hub_offline_online_split": "main_frame.py 27077 -> 27107 (+30): open_speech_hub() now splits its Read Aloud engine_options/engine_available and Dictation all_providers into Offline/Online halves (four kwargs dicts instead of two) for the Speech Hub's new 4-tab split (#847); choose_read_aloud_configuration()/the set_default dictation-result handler add a few lines each. The heavier logic (engine_scope filtering, the 4-tab notebook, the no-cloud-provider empty state) lives in speech_setup_dialog.py and speech_hub_dialog.py, both still well under the 600-line default cap.",
     "_rebaseline_2026_07_08_wizard_data_location_restart_offer": "main_frame.py 27107 -> 27117 (+10): run_startup_wizard() now offers the same Restart Now/Later prompt Preferences already gives (_confirm_restart_for_data_location) when the wizard's Data Location page queues a real move -- a first-run choice to store data elsewhere otherwise looked silently ignored until the second launch, since the move is deferred to next launch the same way Preferences' change is (not safe to apply live). Uses the new quill.core.data_location.pending_data_location_target().",
     "_rebaseline_2026_07_11_cmd_menu_label_native_accelerator_fix": "main_frame.py 28119 -> 28132 (+13): _menu_label() translates a literal 'Cmd' modifier token to 'Ctrl' before embedding a binding into a menu label's native-parsed tab-accelerator slot -- wx's wxGetAccelFromString does not reliably recognize 'Cmd', silently dropping it and leaving a bare, unmodified key (e.g. Find Next's macOS-only 'Cmd+G' left bare G bound as a real accelerator, intercepting every G/g keystroke before it reached the editor). Same bug class as #612 (QUILL Key chords), different unrecognized token.",
-    "quill/ui/main_frame.py": 19212,
+    "quill/ui/main_frame.py": 19216,
+    "_rebaseline_2026_07_16_speech_mixins_wiring": "main_frame.py 19212 -> 19216 (+4): two imports and two base-class entries for SpeechDownloadsMixin and VoiceInteractionMixin, the two mixins split out of SpeechCommandsMixin this session. Pure MRO wiring; the moved code lives in main_frame_speech_downloads.py / main_frame_speech_voice.py.",
     "_rebaseline_2026_07_16_cq1_mixin_extraction_wave": "CQ-1 decomposition wave: main_frame.py ratcheted 28261 -> 19212 (-9049, 32%) by extracting ten cohesive clusters verbatim into new mixin modules, each tracked below: main_frame_commands.py 2450 (the whole _build_commands registry + _command_to_menu_id_map), main_frame_preferences.py 1521 (Preferences hub, spec-driven General Preferences, GLOW/AI settings, apply-refresh, data-location restart, legacy import, and _confirm_show_editor_border relocated next to its only caller), main_frame_updates.py 908 (self-update + GLOW update flows), main_frame_spellcheck.py 734 (spelling review, misspelling nav, thesaurus), main_frame_bw.py 708 (dictation toggle + BITS Whisperer provider/model management), main_frame_menu_editor.py 635 (Menu Editor + _TOP_MENU_DEFS/_normalize_menu_label, now imported by main_frame.py), main_frame_search.py 635 (find/replace + search-in-files), main_frame_keymap_io.py 617 (keyboard packs, keymap/share import-export, factory resets); main_frame_compare.py and main_frame_publishing.py landed under the 600-line default cap. Dialog and accessible-name inventories regenerated 1:1 (moved keys only, classifications unchanged).",
     "quill/ui/main_frame_commands.py": 2450,
     "quill/ui/main_frame_preferences.py": 1521,
@@ -505,7 +506,10 @@
     "_rebaseline_2026_07_03_pandoc_download_progress_dialog": "main_frame_speech.py 2097 -> 2126 (+29): download_pandoc rewritten from the silent _run_background_task (status-bar-only feedback) to the AIProgressDialog + threading.Thread pattern used by download_ffmpeg/download_offline_speech_engine/download_faster_whisper, so the Pandoc download (triggered from File > Convert File's first-use prompt, #806) shows a visible, cancelable percentage instead of relying solely on a status-bar line and one-shot announcement that's easy to miss.",
     "_rebaseline_2026_07_08_hub_reopen_preselect": "main_frame_speech.py 2775 -> 2778 (+3): the hub's _back() closure and the guided offline-speech _done() on_ok both now call open_optional_components(preselect=...) instead of no-arg, so returning from a download reselects the row just downloaded instead of jarringly resetting to the top of the list.",
     "_rebaseline_2026_07_08_hub_voice_preview_stop_button": "hub voice-preview Test<->Stop button feedback. main_frame_speech.py 2792 -> 2816 (+24): ComponentsController.test()/_test_optional_component() gained on_state_change forwarding to the hub's Test<->Stop button toggle, so users see real-time feedback during voice preview playback.",
-    "quill/ui/main_frame_speech.py": 2931,
+    "quill/ui/main_frame_speech.py": 462,
+    "_rebaseline_2026_07_16_speech_decomposition": "CQ-1: SpeechCommandsMixin split; main_frame_speech.py 2608 -> 462 (budget was a stale 2931). The model/component download surface (Speech Models manager, HuggingFace token, Optional Components hub, guided setup, and every download_* acquisition) moved verbatim into SpeechDownloadsMixin (main_frame_speech_downloads.py, 1784); the offline voice-interaction surface (dictation, voice commands, conversation mode, Hey QUILL wake word) moved into VoiceInteractionMixin (main_frame_speech_voice.py, 724). The core file keeps the provider accessors, transcription, captions, microphone selection, and command registration. Source-contract tests that grep main_frame_speech.py now scan all three main_frame_speech*.py files.",
+    "quill/ui/main_frame_speech_downloads.py": 1784,
+    "quill/ui/main_frame_speech_voice.py": 724,
     "_rebaseline_2026_07_14_adp_hooks": "Pre-release ADP Assistant hooks (unlock-gated, undocumented by request). main_frame_menu.py 5094 -> 5095 (+1: the _append_adp_media_items call in the Media submenu builder). main_frame_speech.py 2900 -> 2905 (+5: the voice-question router tries the ADP route before Ask Quill; the route itself lives in main_frame_adp.py). The feature bodies are in the new quill/core/adp/ and quill/ui/adp_dialog.py / main_frame_adp.py modules.",
     "_rebaseline_2026_07_14_braille_pack_menu_rebuild_fix": "main_frame_speech.py 2897 -> 2900 (+3): download_braille_pack()'s success callback called self._request_menu_refresh() (which only refreshes existing menu-item state) instead of self._build_menu() (which structurally rebuilds submenus), so the Tools > Braille > Translation submenu never appeared after a successful download without a restart -- same root cause class as #974 (Quillin menu contributions).",
     "_rebaseline_2026_07_10_macos_batch3a": "macOS review batch 3a. main_frame_speech.py 2879 -> 2894 (+15): download_braille_pack gains a braille_install_supported() gate that shows a 'Windows-only, use Homebrew liblouis' message on macOS instead of attempting the Windows-binary download (#47). structured.py 702 -> 712 (+10): the two LibreOffice conversion helpers now resolve soffice via the new quill.core.external_tools.libreoffice_executable() (which finds the macOS .app install), instead of the bare 'soffice' argv that missed every standard Mac (#41). Both are small, single-purpose growths in CQ-1 extraction candidates; the macOS resolution helpers themselves live in the wx-free quill/core/external_tools.py and quill/core/braille_pack.py.",
