+++ tests/unit/ui/fixtures/dialog_inventory.json
@@ -357,13 +357,13 @@
   "quill/ui/main_frame_sessions.py::SessionsMixin.save_session::wx.FileDialog": "native",
   "quill/ui/main_frame_simple_open.py::SimpleOpenMixin._prompt_native_open_dialog::wx.FileDialog": "native",
   "quill/ui/main_frame_speech.py::SpeechCommandsMixin._choose_transcript_format::wx.SingleChoiceDialog": "native",
-  "quill/ui/main_frame_speech.py::SpeechCommandsMixin._choose_voice_to_preview::wx.SingleChoiceDialog": "native",
   "quill/ui/main_frame_speech.py::SpeechCommandsMixin._save_captions::wx.FileDialog": "native",
   "quill/ui/main_frame_speech.py::SpeechCommandsMixin._save_captions::wx.SingleChoiceDialog": "native",
   "quill/ui/main_frame_speech.py::SpeechCommandsMixin.choose_dictation_microphone::wx.SingleChoiceDialog": "native",
   "quill/ui/main_frame_speech.py::SpeechCommandsMixin.generate_captions_offline::wx.FileDialog": "native",
-  "quill/ui/main_frame_speech.py::SpeechCommandsMixin.set_huggingface_token::wx.PasswordEntryDialog": "native",
   "quill/ui/main_frame_speech.py::SpeechCommandsMixin.transcribe_audio_offline::wx.FileDialog": "native",
+  "quill/ui/main_frame_speech_downloads.py::SpeechDownloadsMixin._choose_voice_to_preview::wx.SingleChoiceDialog": "native",
+  "quill/ui/main_frame_speech_downloads.py::SpeechDownloadsMixin.set_huggingface_token::wx.PasswordEntryDialog": "native",
   "quill/ui/main_frame_spellcheck.py::SpellcheckCommandsMixin._choose_misspelling_with_context::wx.Dialog": "hardened_custom",
   "quill/ui/main_frame_spellcheck.py::SpellcheckCommandsMixin.show_thesaurus::wx.SingleChoiceDialog": "native",
   "quill/ui/main_frame_spellcheck.py::SpellcheckCommandsMixin.show_thesaurus::wx.SingleChoiceDialog#2": "native",
